Setting MEAN and VARIANCE after initialization
I suggest to add setters function for MEAN and VARIANCE. Reason behind this is a need to work with different distribution object in a loop. If we don't have these setters function, then we have to create an object of Normal class in each iteration of loop passing its Mean and Variance in the constructor. Instead, we can create only one object outside the loop and keep setting MEAN and VARIANCE in each iteration for every different distribution.